Move cycle counter to model
